[MIG] stock_account_operating_unit: Migration to 10.0
