+++ stock_account_operating_unit/__manifest__.py
@@ -8,13 +8,13 @@
     "name": "Stock account moves with Operating Unit",
     "summary": "Create journal entries in moves between internal locations "
                "with different operating units.",
-    "version": "9.0.1.0.0",
+    "version": "10.0.1.0.0",
     "category": "Generic Modules/Sales & Purchases",
     "author": "Eficent Business and IT Consulting Services S.L., "
               "Serpent Consulting Services Pvt. Ltd.,"
               "Odoo Community Association (OCA)",
     "license": "LGPL-3",
-    "website": "http://www.eficent.com",
+    "website": "https://github.com/OCA/operating-unit",
     "depends": [
         'stock_operating_unit',
         'account_operating_unit',
