+++ stock_account_operating_unit/model/__init__.py
@@ -3,4 +3,5 @@
 # - Jordi Ballester Alomar
 # © 2015-17 Serpent Consulting Services Pvt. Ltd. - Sudhir Arya
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
-from . import stock
+from . import stock_move
+from . import stock_quant
